refactor: 중복된 도메인 변경
