+++ src/main/java/hello/servlet/ServletApplication.java
@@ -1,4 +1,4 @@
-package hello.servlet.servelt;
+package hello.servlet;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
