+++ src/main/java/hello/servlet/ServletInitializer.java
@@ -1,4 +1,4 @@
-package hello.servlet.servelt;
+package hello.servlet;
 
 import org.springframework.boot.builder.SpringApplicationBuilder;
 import org.springframework.boot.web.servlet.support.SpringBootServletInitializer;
