+++ src/main/java/hello/servlet/basic/HelloServlet.java
@@ -1,4 +1,4 @@
-package hello.servlet.servelt.basic;
+package hello.servlet.basic;
 
 import jakarta.servlet.ServletException;
 import jakarta.servlet.ServletRequest;
