+++ src/main/java/hello/servlet/basic/request/RequestBodyJsonServlet.java
@@ -1,7 +1,7 @@
-package hello.servlet.servelt.basic.request;
+package hello.servlet.basic.request;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import hello.servlet.servelt.basic.HelloData;
+import hello.servlet.basic.HelloData;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.ServletInputStream;
 import jakarta.servlet.annotation.WebServlet;
