+++ src/main/java/hello/servlet/basic/request/RequestHeaderServlet.java
@@ -1,4 +1,4 @@
-package hello.servlet.servelt.basic.request;
+package hello.servlet.basic.request;
 
 import jakarta.servlet.ServletException;
 import jakarta.servlet.annotation.WebServlet;
