+++ src/main/java/hello/servlet/basic/request/RequestParamServlet.java
@@ -1,4 +1,4 @@
-package hello.servlet.servelt.basic.request;
+package hello.servlet.basic.request;
 
 import jakarta.servlet.ServletException;
 import jakarta.servlet.annotation.WebServlet;
@@ -7,7 +7,6 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 
 import java.io.IOException;
-import java.util.Enumeration;
 
 /**
  * 1. 파라미터 전송 긴능
