+++ src/main/java/hello/servlet/basic/response/ResponseHeaderServlet.java
@@ -1,4 +1,4 @@
-package hello.servlet.servelt.basic.response;
+package hello.servlet.basic.response;
 
 import jakarta.servlet.ServletException;
 import jakarta.servlet.annotation.WebServlet;
